better way of representing sequences of different lengths
All of the outputs at a timestep are stored in a single anyvec.Vector, which is desirable for performance reasons (especially for RNNs). Compare this to autofunc/seqfunc, which represents each sequence as its own []linalg.Vector. In autofunc/seqfunc, a different chunk of memory is responible for each output of each sequence at each timestep. To do any batch operations on the structure in autofunc/seqfunc, the vectors at a timestep would first need to be concatenated before being operated on.

In general, I think using a packed array of vectors combined with a Present mapping is a much more robust solution.