sequence pooling
I was debating with myself a bit about how to do this. The current approach--also the simplest--creates a separate pooling Var per timestep. The problem with this is that now everytihng that uses the seq will depend on a boatload of Vars. This in turn means dealing with fairly large maps (anydiff.VarSet objects). However, I do not think this will end up being a performance bottleneck for most applications.